requirements.txt
boto3==1.26.0
pandas==1.5.0
PyPDF2==3.0.0
python-dotenv==0.21.0
redis==4.3.4